Implementing default user picture. But a OneToOne relationship with files doesn't seem right for this.

+++ srcs/requirements/backend/app/src/modules/users/services/pictures/pictures.service.ts
@@ -1,8 +1,14 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException, UnprocessableEntityException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from '../../entities/User.entity';
 import { Equal, Repository } from 'typeorm';
 import { File } from '../../../file-uploads/entities/file.entity';
+import { DEFAULT_USER_PICTURE_PATH } from 'src/modules/file-uploads/configs/file-uploads.constants';
+import * as fs from 'fs';
+import * as chardet from 'chardet';
+import * as path from 'path';
+import * as mimeTypes from 'mime-types';
+import { MimeTypes } from 'src/modules/file-uploads/enums/mime-types.enum';
 
 @Injectable()
 export class PicturesService {
@@ -41,13 +47,12 @@ export class PicturesService {
 
         if (user.picture) throw new BadRequestException(`User ${username ? username : '{undefined}'} already has a picture`);
 
-        console.log(file);
-
         user.picture = this.fileRepository.create({
             ...file,
         });
 
         await this.userRepository.save(user);
+
         return (user.picture);
     }
 
@@ -68,9 +73,10 @@ export class PicturesService {
         user.picture = this.fileRepository.create({
             ...file,
         });
-        await this.userRepository.save(user);
 
+        await this.userRepository.save(user);
         await this.fileRepository.remove(picture);
+
         return (user.picture);
     }
 
@@ -86,12 +92,70 @@ export class PicturesService {
 
         if (!picture) throw new BadRequestException(`User ${username ? username : '{undefined}'} has no picture`);
         
-        user.picture = null;
-        await this.userRepository.save(user);
+        const defaultPicture = await this.getDefaultUserPicture();
 
+        if (picture.id === defaultPicture.id) throw new BadRequestException(`Default user picture already set`);
+
+        user.picture = defaultPicture;
+        await this.userRepository.save(user);
         await this.fileRepository.remove(picture);
 
         return (`User '${username}'s picture successfully deleted`);
+    }
+
+    /* Default user picture */
+
+    async getDefaultUserPicture(): Promise<File> {
+        const picture = await this.fileRepository.findOne({
+            where: { fieldname: Equal('default_user_picture') },
+        });
+
+        if (!picture) throw new NotFoundException(`User default picture not found`);
+
+        return (picture);
+    }
+
+    async createDefaultUserPicture(): Promise<void> {
+        let picture = await this.fileRepository.findOne({
+            where: { fieldname: Equal('default_user_picture') },
+        });
+
+        if (picture) throw new BadRequestException(`Default user picture already exists as File with ID ${picture.id}`);
+
+        const defaultUserPicturePath = DEFAULT_USER_PICTURE_PATH;
+        let size: number;
+
+        try {
+            size = fs.statSync(defaultUserPicturePath).size;
+        } catch(error) {
+            throw new UnprocessableEntityException(`Couldn't read ${defaultUserPicturePath}. Verify if file exists`);
+        }
+
+        const filename = path.basename(defaultUserPicturePath);
+        const encoding = this.detectEncoding(defaultUserPicturePath);
+        const mimetype = this.detectMimetype(defaultUserPicturePath);
+
+        picture = this.fileRepository.create({
+            fieldname: 'default_user_picture',
+            originalname: filename,
+            encoding: encoding,
+            mimetype: mimetype,
+            size: size,
+            path: defaultUserPicturePath,
+        });
+
+        await this.fileRepository.save(picture);
+    }
+
+    private detectEncoding(filePath: string): string {
+        const fileData = fs.readFileSync(filePath);
+        return (chardet.detect(fileData));
+    }
+
+    private detectMimetype(filePath: string): MimeTypes {
+        const mimeType = mimeTypes.lookup(filePath);
+        if (!mimeType || !Object.values(MimeTypes).includes(mimeType as MimeTypes)) throw new UnprocessableEntityException(`MimeType ${mimeType} not recognized`);
+        return (mimeType as MimeTypes);
     }
 
 }
